test: separate mui barrel optimization test
