+++ test/development/basic/barrel-optimization/barrel-optimization-mui.test.ts
@@ -1,0 +1,46 @@
+import { join } from 'path'
+import { createNextDescribe } from 'e2e-utils'
+import { hasRedbox } from 'next-test-utils'
+// Skipped in Turbopack, will be added later.
+;(process.env.TURBOPACK ? describe.skip : describe)(
+  'Skipped in Turbopack',
+  () => {
+    createNextDescribe(
+      'optimizePackageImports - mui',
+      {
+        env: {
+          NEXT_TEST_MODE: '1',
+        },
+        files: join(__dirname, 'fixture'),
+
+        dependencies: {
+          '@mui/material': '5.15.4',
+        },
+      },
+      ({ next }) => {
+        it('should support MUI', async () => {
+          let logs = ''
+          next.on('stdout', (log) => {
+            logs += log
+          })
+
+          // Ensure that MUI is working
+          const $ = await next.render$('/mui')
+          expect(await $('#button').text()).toContain('button')
+          expect(await $('#typography').text()).toContain('typography')
+
+          const browser = await next.browser('/mui')
+          expect(await hasRedbox(browser)).toBe(false)
+
+          const modules = [...logs.matchAll(/\((\d+) modules\)/g)]
+          expect(modules.length).toBeGreaterThanOrEqual(1)
+          for (const [, moduleCount] of modules) {
+            // Ensure that the number of modules is less than 1500 - otherwise we're
+            // importing the entire library.
+            expect(parseInt(moduleCount)).toBeLessThan(1500)
+          }
+        })
+      }
+    )
+  }
+)
