+++ test/development/basic/barrel-optimization/barrel-optimization.test.ts
@@ -1,36 +1,22 @@
 import { join } from 'path'
 import { createNextDescribe } from 'e2e-utils'
-import { hasRedbox, shouldRunTurboDevTest } from 'next-test-utils'
 // Skipped in Turbopack, will be added later.
 ;(process.env.TURBOPACK ? describe.skip : describe)(
   'Skipped in Turbopack',
   () => {
     createNextDescribe(
-      'optimizePackageImports',
+      'optimizePackageImports - basic',
       {
         env: {
           NEXT_TEST_MODE: '1',
         },
         files: join(__dirname, 'fixture'),
-        packageJson: {
-          scripts: {
-            build: `next build`,
-            dev: `next ${shouldRunTurboDevTest() ? 'dev --turbo' : 'dev'}`,
-            start: 'next start',
-          },
-        },
-        installCommand: 'pnpm i',
-        startCommand: (global as any).isNextDev
-          ? 'pnpm run dev'
-          : 'pnpm run start',
-        buildCommand: 'pnpm run build',
         dependencies: {
           'lucide-react': '0.264.0',
           '@headlessui/react': '1.7.17',
           '@heroicons/react': '2.0.18',
           '@visx/visx': '3.3.0',
           'recursive-barrel': '1.0.0',
-          '@mui/material': '5.15.4',
           '@emotion/styled': '11.11.0',
           '@emotion/react': '11.11.1',
         },
@@ -130,29 +116,6 @@ import { hasRedbox, shouldRunTurboDevTest } from 'next-test-utils'
         it('should support visx', async () => {
           const html = await next.render('/visx')
           expect(html).toContain('<linearGradient')
-        })
-
-        it('should support MUI', async () => {
-          let logs = ''
-          next.on('stdout', (log) => {
-            logs += log
-          })
-
-          // Ensure that MUI is working
-          const $ = await next.render$('/mui')
-          expect(await $('#button').text()).toContain('button')
-          expect(await $('#typography').text()).toContain('typography')
-
-          const browser = await next.browser('/mui')
-          expect(await hasRedbox(browser)).toBe(false)
-
-          const modules = [...logs.matchAll(/\((\d+) modules\)/g)]
-          expect(modules.length).toBeGreaterThanOrEqual(1)
-          for (const [, moduleCount] of modules) {
-            // Ensure that the number of modules is less than 1500 - otherwise we're
-            // importing the entire library.
-            expect(parseInt(moduleCount)).toBeLessThan(1500)
-          }
         })
 
         it('should not break "use client" directive in optimized packages', async () => {
